requirements.txt
flask==3.1.3
flask-cors==6.0.2
google-genai==1.64.0
supabase==2.28.0
python-dotenv==1.2.1
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python-headless>=4.8.0
